Trait pallet_moonbeam_lazy_migrations::pallet::Config

pub trait Config: Config + Config + Config + Config<AssetId = u128> + Config<AssetId = u128> + Config {
    type ForeignAssetMigratorOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configuration trait of this pallet.

Required Associated Types

type ForeignAssetMigratorOrigin: EnsureOrigin<Self::RuntimeOrigin>

type WeightInfo: WeightInfo

Object Safety

This trait is not object safe.